Matches-Today
This Python script allows you to easily retrieve a list of football (soccer) matches that are being played on the current day.
The script utilizes online resources to gather the necessary information and presents it in a user-friendly format. 
making it an ideal tool for fans who want to stay up-to-date with their favorite teams and leagues. 
The script is available on GitHub for you to download and use as you see fit. Whether you are a passionate football fan or simply want to keep track of the latest matches, this script is a convenient and reliable resource.